refactor(server)!: give expires_at and auth_time one type each
`expires_at` was a `timestamp` on `oidc_sessions` and
`oidc_authentication_contexts` but a `dateTime` on the three token tables;
`auth_time` was an `unsignedInteger` on contexts and an `unsignedBigInteger`
on authorization codes. Both now use the wider of the two everywhere.

`AuthorizationCode` casts `auth_time` as `integer` rather than `int`, matching
`AuthenticationContext`.

+++ packages/server/database/migrations/2026_07_16_000001_create_oidc_authentication_contexts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->uuid('session_id')->nullable()->index();
             $table->json('amr');
             $table->string('acr')->nullable();
-            $table->unsignedInteger('auth_time')->nullable();
+            $table->unsignedBigInteger('auth_time')->nullable();
             $table->json('id_token_claims');
             $table->json('access_token_claims');
             $table->timestamp('created_at')->nullable();
